backend/requirements.txt
# requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
whisper==1.0.0
torch==1.9.0
transformers==4.11.3
python-dotenv==0.19.0
pytest==6.2.5